server/requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
boto3==1.20.18
botocore==1.23.18
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-apscheduler==0.6.0
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-environ==0.7.0
django-extensions==3.1.3
django-filter==21.1
django-js-asset==1.2.2
django-ses==2.3.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
environ==1.0
future==0.18.2
google-api-core==2.1.1
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.4.0
protobuf==3.18.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
